+++ crates/wasm-encoder/src/component/exports.rs
@@ -1,5 +1,8 @@
-use super::{COMPONENT_SORT, CORE_SORT, FUNCTION_SORT, INSTANCE_SORT, TYPE_SORT, VALUE_SORT};
-use crate::{encode_section, ComponentSection, ComponentSectionId, Encode, CORE_MODULE_SORT};
+use super::{
+    COMPONENT_SORT, CORE_MODULE_SORT, CORE_SORT, FUNCTION_SORT, INSTANCE_SORT, TYPE_SORT,
+    VALUE_SORT,
+};
+use crate::{encode_section, ComponentSection, ComponentSectionId, Encode};
 
 /// Represents the kind of an export from a WebAssembly component.
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
@@ -44,66 +47,16 @@ impl Encode for ComponentExportKind {
     }
 }
 
-/// Represents an export from a WebAssembly component.
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum ComponentExport {
-    /// The export is a core module.
-    Module(u32),
-    /// The export is a function.
-    Func(u32),
-    /// The export is a value.
-    Value(u32),
-    /// The export is a type.
-    Type(u32),
-    /// The export is an instance.
-    Instance(u32),
-    /// The export is a component.
-    Component(u32),
-}
-
-impl ComponentExport {
-    /// Gets the kind of the export.
-    pub fn kind(&self) -> ComponentExportKind {
-        match self {
-            Self::Module(_) => ComponentExportKind::Module,
-            Self::Func(_) => ComponentExportKind::Func,
-            Self::Value(_) => ComponentExportKind::Value,
-            Self::Type(_) => ComponentExportKind::Type,
-            Self::Instance(_) => ComponentExportKind::Instance,
-            Self::Component(_) => ComponentExportKind::Component,
-        }
-    }
-
-    /// Gets the index of the export.
-    fn index(&self) -> u32 {
-        match self {
-            Self::Module(idx)
-            | Self::Func(idx)
-            | Self::Value(idx)
-            | Self::Type(idx)
-            | Self::Instance(idx)
-            | Self::Component(idx) => *idx,
-        }
-    }
-}
-
-impl Encode for ComponentExport {
-    fn encode(&self, sink: &mut Vec<u8>) {
-        self.kind().encode(sink);
-        self.index().encode(sink);
-    }
-}
-
 /// An encoder for the export section of WebAssembly component.
 ///
 /// # Example
 ///
 /// ```rust
-/// use wasm_encoder::{Component, ComponentExportSection, ComponentExport};
+/// use wasm_encoder::{Component, ComponentExportSection, ComponentExportKind};
 ///
 /// // This exports a function named "foo"
 /// let mut exports = ComponentExportSection::new();
-/// exports.export("foo", ComponentExport::Func(0));
+/// exports.export("foo", ComponentExportKind::Func, 0);
 ///
 /// let mut component = Component::new();
 /// component.section(&exports);
@@ -133,9 +86,10 @@ impl ComponentExportSection {
     }
 
     /// Define an export in the export section.
-    pub fn export(&mut self, name: &str, export: ComponentExport) -> &mut Self {
+    pub fn export(&mut self, name: &str, kind: ComponentExportKind, index: u32) -> &mut Self {
         name.encode(&mut self.bytes);
-        export.encode(&mut self.bytes);
+        kind.encode(&mut self.bytes);
+        index.encode(&mut self.bytes);
         self.num_added += 1;
         self
     }
