+++ crates/wasm-encoder/src/core.rs
@@ -1,4 +1,3 @@
-mod aliases;
 mod code;
 mod custom;
 mod data;
@@ -7,7 +6,6 @@ mod exports;
 mod functions;
 mod globals;
 mod imports;
-mod instances;
 mod linking;
 mod memories;
 mod names;
@@ -16,7 +14,6 @@ mod tables;
 mod tags;
 mod types;
 
-pub use aliases::*;
 pub use code::*;
 pub use custom::*;
 pub use data::*;
@@ -25,7 +22,6 @@ pub use exports::*;
 pub use functions::*;
 pub use globals::*;
 pub use imports::*;
-pub use instances::*;
 pub use linking::*;
 pub use memories::*;
 pub use names::*;
@@ -41,9 +37,6 @@ pub(crate) const CORE_TABLE_SORT: u8 = 0x01;
 pub(crate) const CORE_MEMORY_SORT: u8 = 0x02;
 pub(crate) const CORE_GLOBAL_SORT: u8 = 0x03;
 pub(crate) const CORE_TAG_SORT: u8 = 0x04;
-pub(crate) const CORE_TYPE_SORT: u8 = 0x10;
-pub(crate) const CORE_MODULE_SORT: u8 = 0x11;
-pub(crate) const CORE_INSTANCE_SORT: u8 = 0x12;
 
 /// A WebAssembly module section.
 ///
